AI-Model/requirements.txt

flask
transformers
torch
keybert
sentence-transformers
Pillow
requests
